+++ ConsoleApplication3/olcPixelGameEngine.cpp
@@ -1,0 +1,2 @@
+#define OLC_PGE_APPLICATION
+#include "olcPixelGameEngine.h"
